More graph code re-use
