+++ cpp/dolfinx/graph/partitioners.h
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 Igor A. Baratta
+// Copyright (C) 2020 Garth N. Wells
 //
 // This file is part of DOLFINx (https://www.fenicsproject.org)
 //
@@ -6,13 +6,30 @@
 
 #pragma once
 
+#include "partition.h"
 #include <cstdint>
 #include <dolfinx/graph/AdjacencyList.h>
 #include <functional>
 #include <mpi.h>
 
+namespace dolfinx::graph
+{
+
+namespace parmetis
+{
+#ifdef HAS_PARMETIS
+
+/// Create a graph partitioning function that uses ParMETIS
+///
+/// param[in] options The ParMETIS option. See ParMETIS manual for
+/// details.
+graph::partition_fn partitioner(std::array<int, 3> options = {0, 0, 0});
+
+#endif
+} // namespace parmetis
+
 /// Interfaces to KaHIP parallel partitioner
-namespace dolfinx::graph::kahip
+namespace kahip
 {
 #ifdef HAS_KAHIP
 /// Create a graph partitioning function that uses KaHIP
@@ -24,10 +41,10 @@ namespace dolfinx::graph::kahip
 /// @param[in] suppress_output Suppresses KaHIP output if true
 /// @return A KaHIP graph partitioning function with specified parameter
 /// options
-std::function<graph::AdjacencyList<std::int32_t>(
-    MPI_Comm, int, const AdjacencyList<std::int64_t>&, std::int32_t, bool)>
-partitioner(int mode = 1, int seed = 0, double imbalance = 0.03,
-            bool suppress_output = true);
+graph::partition_fn partitioner(int mode = 1, int seed = 0,
+                                double imbalance = 0.03,
+                                bool suppress_output = true);
 
 #endif
-} // namespace dolfinx::graph::kahip
+} // namespace kahip
+} // namespace dolfinx::graph
